:smile: Latest changes

+++ src/tasks/deleteData.ts
@@ -1,6 +1,7 @@
 import { Task, log } from "actionhero";
 const User = require("../Models/User");
 const Message = require("../Models/Message");
+const Room = require("../Models/Room");
 
 export class DeleteData extends Task {
   constructor() {
@@ -18,8 +19,14 @@ export class DeleteData extends Task {
   async run(data) {
     // your logic here
     try {
-      await User.deleteMany({ room_id: data.params.room_id });
-      await Message.deleteMany({ room_id: data.params.room_id });
+      let room = await Room.findOne(data.params.room_id);
+      if (room) {
+        await User.deleteMany({ room_id: data.params.room_id });
+        await Message.deleteMany({ room_id: data.params.room_id });
+        await Room.delete({ room_id: data.params.room_id });
+      } else {
+        throw Error;
+      }
     } catch (e) {
       log(`An error occured whilst deleting data: ${e.error}`, "error");
       throw new Error(e);
